+++ .ipynb_checkpoints/README-checkpoint.md
@@ -1,0 +1,33 @@
+
+# Implementing Generative AI on AWS workshop
+
+
+This workshop is set up following the popular AWS Immersion Day format. It means to provide guidance on how to get started with Generative AI on AWS. The Immersion Day is split up into the following four blocks, consisting of a theory section covered by slides as well of a hands-on lab each:
+- Introduction Generative AI & Large Language Models, Large Language Model deployment & inference optimization
+- Introduction Visual Foundation Models, deployment & inference optimization of Stable Diffusion
+- Large Language Model finetuning
+- Engineering GenAI-powered applications on AWS
+
+Note that during an immersion day / workshop potentially only a subset of these topics might be covered.
+
+The repository is structured as follows: The slides can be found in the GenerativeAIImmersionDayPresentationDeck.pdf residing on root level of the repository. Similarily, the labs can be found in respectively named directories: 
+- Lab 1 - Hosting Large Language Models can be found in the lab1 directory. 
+    - Option 1: For GPT-J start with the notebook option-1-gpt-j-notebook-full.ipynb.
+    - Option 2: For Falcon7b-instruct start with the notebook falcon7b-instruct-notebook-full.ipynb.
+    - Option 3: For Falcon40b-instruct start with the notebook falcon40b-instruct-notebook-full.ipynb.
+- Lab 2 - Hosting Stable Diffusion can be found in the lab3 directory. Start with the notebook JumpStart_Stable_Diffusion_Inference_Only.ipynb.
+- Lab (optional) - Finetuning Large Language Models can be found in the lab2 directory. Start with the notebook fine-tuning.ipynb.
+
+## Security
+
+See [CONTRIBUTING](CONTRIBUTING.md#security-issue-notifications) for more information.
+
+## License
+
+This library is licensed under the MIT-0 License. See the LICENSE file.
+
+
+## References
+This workshop is an adaptation of the given full workshop Generative AI on AWS workshop present on:  https://catalog.us-east-1.prod.workshops.aws/workshops/80ae1ed2-f415-4d3d-9eb0-e9118c147bd4
+
+
