--- lab2/finetuning/requirements.txt
@@ -1,8 +1,0 @@
-accelerate
-torch >= 1.3
-datasets == 2.10.1
-sentencepiece != 0.1.92
-protobuf
-scikit-learn
-transformers==4.17.0
-evaluate==0.4.0
