--- lab4/rag_app/kendra/__init__.py
@@ -1,1 +1,0 @@
-"""Classes to work with AWS Kendra and Bedrock LLMs"""
