--- lab4/rag_app/requirements.txt
@@ -1,2 +1,0 @@
-boto3>=1.26.163
-langchain>=0.0.219
